Struct twilight_embed_builder::builder::EmbedBuilder

#[must_use = "must be built into an embed"]pub struct EmbedBuilder(_);

Create an embed with a builder.

Examples

Refer to the crate-level documentation for examples.

Implementations

impl EmbedBuilder

pub const COLOR_MAXIMUM: u32

The maximum accepted color value.

This is used by color.

pub const DESCRIPTION_LENGTH_LIMIT: usize

The maximum number of UTF-16 code points that can be in a description.

This is used by description.

pub const EMBED_FIELD_LIMIT: usize

The maximum number of fields that can be in an embed.

This is used by build.

pub const EMBED_LENGTH_LIMIT: usize

The maximum total textual length of the embed in UTF-16 code points.

This combines the text of the author name, description, footer text, field names and values, and title.

This is used by build.

pub const TITLE_LENGTH_LIMIT: usize

The maximum number of UTF-16 code points that can be in a title.

This is used by title.

pub fn new() -> Self

Create a new default embed builder.

See the crate-level documentation for examples and additional information.

This is equivalent to the default implementation.

#[must_use = "should be used as part of something like a message"]pub fn build(self) -> Result<Embed, EmbedBuildError>

Build this into an embed.

Errors

Returns EmbedBuildError::ContentTooLarge if the textual content of the embed is too large. Refer to EMBED_LENGTH_LIMIT for the limit value and what counts towards it.

Returns EmbedBuildError::TooManyFields if there are too many fields in the embed. Refer to EMBED_FIELD_LIMIT for the limit value.

pub fn author(self, author: impl Into<EmbedAuthor>) -> Self

Set the author.

Examples

Create an embed author:

use twilight_embed_builder::{EmbedAuthorBuilder, EmbedBuilder};

let author = EmbedAuthorBuilder::new()
    .name("Twilight")?
    .url("https://github.com/twilight-rs/twilight")
    .build();

let embed = EmbedBuilder::new().author(author).build();

pub fn color(self, color: u32) -> Result<Self, EmbedColorError>

Set the color.

This must be a valid hexadecimal RGB value. 0x000000 is not an acceptable value as it would be thrown out by Discord.

Examples

Set the color of an embed to 0xfd69b3:

use twilight_embed_builder::EmbedBuilder;

let embed = EmbedBuilder::new()
    .color(0xfd_69_b3)?
    .description("a description")?
    .build();

Errors

Returns EmbedError::NotRgb if the provided color is not a valid RGB integer. Refer to COLOR_MAXIMUM to know what the maximum accepted value is.

Returns EmbedError::Zero if the provided color is 0, which is not an acceptable value.

pub fn description(
self,
description: impl Into<String>
) -> Result<Self, EmbedDescriptionError>

Set the description.

Refer to DESCRIPTION_LENGTH_LIMIT for the maximum number of UTF-16 code points that can be in a description.

Examples

use twilight_embed_builder::EmbedBuilder;

let embed = EmbedBuilder::new().description("this is an embed")?.build();

Errors

Returns EmbedDescriptionError::DescriptionTooLong if the provided description is longer than the maximum number of code points.

pub fn field(self, field: impl Into<EmbedField>) -> Self

Add a field to the embed.

Examples

use twilight_embed_builder::{EmbedBuilder, EmbedFieldBuilder};

let embed = EmbedBuilder::new()
    .description("this is an embed")?
    .field(EmbedFieldBuilder::new("a field", "and its value")?)
    .build()?;

pub fn footer(self, footer: impl Into<EmbedFooter>) -> Self

Set the footer of the embed.

Examples

use twilight_embed_builder::{EmbedBuilder, EmbedFooterBuilder};

let embed = EmbedBuilder::new()
    .description("this is an embed")?
    .footer(EmbedFooterBuilder::new("a footer")?)
    .build()?;

pub fn image(self, image_source: ImageSource) -> Self

Set the image.

Examples

Set the image source to a URL:

use twilight_embed_builder::{EmbedBuilder, EmbedFooterBuilder, ImageSource};

let source = ImageSource::url("https://raw.githubusercontent.com/twilight-rs/twilight/trunk/logo.png")?;
let embed = EmbedBuilder::new()
    .footer(EmbedFooterBuilder::new("twilight")?)
    .image(source)
    .build()?;

pub fn thumbnail(self, image_source: ImageSource) -> Self

Add a thumbnail.

Examples

Set the thumbnail to an image attachment with the filename "twilight.png":

use twilight_embed_builder::{EmbedBuilder, ImageSource};

let embed = EmbedBuilder::new()
    .description("a picture of twilight")?
    .thumbnail(ImageSource::attachment("twilight.png")?)
    .build()?;

pub fn timestamp(self, timestamp: impl Into<String>) -> Self

Set the ISO 8601 timestamp.

pub fn title(self, title: impl Into<String>) -> Result<Self, EmbedTitleError>

Set the title.

Refer to TITLE_LENGTH_LIMIT for the maximum number of UTF-16 code points that can be in a title.

Examples

Set the title to "twilight":

use twilight_embed_builder::EmbedBuilder;

let embed = EmbedBuilder::new()
    .title("twilight")?
    .url("https://github.com/twilight-rs/twilight")
    .build()?;

Errors

Returns EmbedTitleError::Empty if the provided title is empty.

Returns EmbedTitleError::TooLong if the provided title is longer than the limit defined at TITLE_LENGTH_LIMIT.

pub fn url(self, url: impl Into<String>) -> Self

Set the URL.

Examples

Set the URL to twilight's repository:

use twilight_embed_builder::{EmbedBuilder, EmbedFooterBuilder};

let embed = EmbedBuilder::new()
    .description("twilight's repository")?
    .url("https://github.com/twilight-rs/twilight")
    .build()?;