Crate twilight_embed_builder

twilight-embed-builder

twilight-embed-builder is a set of builder for the twilight-rs ecosystem to create a message embed, useful when creating or updating messages.

Examples

Build a simple embed:

use twilight_embed_builder::{EmbedBuilder, EmbedFieldBuilder};

let embed = EmbedBuilder::new()
    .description("Here's a list of reasons why Twilight is the best pony:")
    .field(EmbedFieldBuilder::new("Wings", "She has wings.").inline())
    .field(EmbedFieldBuilder::new("Horn", "She can do magic, and she's really good at it.").inline())
    .build()?;

Build an embed with an image:

use twilight_embed_builder::{EmbedBuilder, ImageSource};

let embed = EmbedBuilder::new()
    .description("Here's a cool image of Twilight Sparkle")
    .image(ImageSource::attachment("bestpony.png")?)
    .build()?;

Re-exports

pub use self::image_source::ImageSource;

Modules

image_source

Sources to image URLs and attachments.

Structs

EmbedAuthorBuilder

Create an embed author with a builder.

EmbedBuilder

Create an embed with a builder.

EmbedError

Error building an embed.

EmbedFieldBuilder

Create an embed field with a builder.

EmbedFooterBuilder

Create an embed footer with a builder.

Enums

EmbedErrorType

Type of EmbedError that occurred.