Enum twilight_embed_builder::EmbedErrorType

#[non_exhaustive]
pub enum EmbedErrorType {

    AuthorNameEmpty {
        name: String,
    },
    AuthorNameTooLong {
        name: String,
    },
    ColorNotRgb {
        color: u32,
    },
    ColorZero,
    DescriptionEmpty {
        description: String,
    },
    DescriptionTooLong {
        description: String,
    },
    FieldNameEmpty {
        name: String,
        value: String,
    },
    FieldNameTooLong {
        name: String,
        value: String,
    },
    FieldValueEmpty {
        name: String,
        value: String,
    },
    FieldValueTooLong {
        name: String,
        value: String,
    },
    FooterTextEmpty {
        text: String,
    },
    FooterTextTooLong {
        text: String,
    },
    TitleEmpty {
        title: String,
    },
    TitleTooLong {
        title: String,
    },
    TotalContentTooLarge {
        length: usize,
    },
    TooManyFields {
        fields: Vec<EmbedField>,
    },

}

Type of EmbedError that occurred.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AuthorNameEmpty

Name is empty.

Fields of AuthorNameEmpty

name: String

Provided name. Although empty, the same owned allocation is included.

AuthorNameTooLong

Name is longer than 256 UTF-16 code points.

Fields of AuthorNameTooLong

name: String

Provided name.

ColorNotRgb

Color was larger than a valid RGB hexadecimal value.

Fields of ColorNotRgb

color: u32

Provided color hex value.

ColorZero

Color was 0. The value would be thrown out by Discord and is equivalent to null.

DescriptionEmpty

Description is empty.

Fields of DescriptionEmpty

description: String

Provided description. Although empty, the same owned allocation is included.

DescriptionTooLong

Description is longer than 2048 UTF-16 code points.

Fields of DescriptionTooLong

description: String

Provided description.

FieldNameEmpty

Name is empty.

Fields of FieldNameEmpty

name: String

Provided name. Although empty, the same owned allocation is included.

value: String

Provided value.

FieldNameTooLong

Name is longer than 256 UTF-16 code points.

Fields of FieldNameTooLong

name: String

Provided name.

value: String

Provided value.

FieldValueEmpty

Value is empty.

Fields of FieldValueEmpty

name: String

Provided name.

value: String

Provided value. Although empty, the same owned allocation is included.

FieldValueTooLong

Value is longer than 1024 UTF-16 code points.

Fields of FieldValueTooLong

name: String

Provided name.

value: String

Provided value.

FooterTextEmpty

Footer text is empty.

Fields of FooterTextEmpty

text: String

Provided text. Although empty, the same owned allocation is included.

FooterTextTooLong

Footer text is longer than 2048 UTF-16 code points.

Fields of FooterTextTooLong

text: String

Provided text.

TitleEmpty

Title is empty.

Fields of TitleEmpty

title: String

Provided title. Although empty, the same owned allocation is included.

TitleTooLong

Title is longer than 256 UTF-16 code points.

Fields of TitleTooLong

title: String

Provided title.

TotalContentTooLarge

The total content of the embed is too large.

Refer to EmbedBuilder::EMBED_LENGTH_LIMIT for more information about what goes into this limit.

Fields of TotalContentTooLarge

length: usize

The total length of the embed.

TooManyFields

Too many fields were provided.

Refer to EmbedBuilder::EMBED_FIELD_LIMIT for more information about what the limit is.

Fields of TooManyFields

fields: Vec<EmbedField>

The provided fields.