Struct twilight_embed_builder::EmbedBuilder

pub struct EmbedBuilder(_);

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Create an embed with a builder.

Examples

Refer to the crate-level documentation for examples.

Implementations

impl EmbedBuilder

pub const AUTHOR_NAME_LENGTH_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum number of UTF-16 code points that can be in an author name.

pub const COLOR_MAXIMUM: u32

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum accepted color value.

pub const DESCRIPTION_LENGTH_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum number of UTF-16 code points that can be in a description.

pub const EMBED_FIELD_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum number of fields that can be in an embed.

pub const EMBED_LENGTH_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum total textual length of the embed in UTF-16 code points.

This combines the text of the author name, description, footer text, field names and values, and title.

pub const FIELD_NAME_LENGTH_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum number of UTF-16 code points that can be in a field name.

pub const FIELD_VALUE_LENGTH_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum number of UTF-16 code points that can be in a field value.

pub const FOOTER_TEXT_LENGTH_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum number of UTF-16 code points that can be in a footer’s text.

pub const TITLE_LENGTH_LIMIT: usize

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

The maximum number of UTF-16 code points that can be in a title.

pub const fn new() -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Create a new default embed builder.

See the crate-level documentation for examples and additional information.

This is equivalent to the default implementation.

pub fn build(self) -> Result<Embed, EmbedError>

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Build this into an embed.

Errors

Returns an EmbedErrorType::AuthorNameEmpty error type if the provided name is empty.

Returns an EmbedErrorType::AuthorNameTooLong error type if the provided name is longer than AUTHOR_NAME_LENGTH_LIMIT.

Returns an EmbedErrorType::ColorNotRgb error type if the provided color is not a valid RGB integer. Refer to COLOR_MAXIMUM to know what the maximum accepted value is.

Returns an EmbedErrorType::ColorZero error type if the provided color is 0, which is not an acceptable value.

Returns an EmbedErrorType::DescriptionEmpty error type if a provided description is empty.

Returns an EmbedErrorType::DescriptionTooLong error type if a provided description is longer than DESCRIPTION_LENGTH_LIMIT.

Returns an EmbedErrorType::FieldNameEmpty error type if a provided field name is empty.

Returns an EmbedErrorType::FieldNameTooLong error type if a provided field name is longer than FIELD_NAME_LENGTH_LIMIT.

Returns an EmbedErrorType::FieldValueEmpty error type if a provided field value is empty.

Returns an EmbedErrorType::FieldValueTooLong error type if a provided field value is longer than FIELD_VALUE_LENGTH_LIMIT.

Returns an EmbedErrorType::FooterTextEmpty error type if the provided text is empty.

Returns an EmbedErrorType::FooterTextTooLong error type if the provided text is longer than the limit defined at FOOTER_TEXT_LENGTH_LIMIT.

Returns an EmbedErrorType::TitleEmpty error type if the provided title is empty.

Returns an EmbedErrorType::TitleTooLong error type if the provided text is longer than the limit defined at TITLE_LENGTH_LIMIT.

Returns an EmbedErrorType::TooManyFields error type if there are too many fields in the embed. Refer to EMBED_FIELD_LIMIT for the limit value.

Returns an EmbedErrorType::TotalContentTooLarge error type if the textual content of the embed is too large. Refer to EMBED_LENGTH_LIMIT for the limit value and what counts towards it.

pub fn author(self, author: impl Into<EmbedAuthor>) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the author.

Examples

Create an embed author:

use twilight_embed_builder::{EmbedAuthorBuilder, EmbedBuilder};

let author = EmbedAuthorBuilder::new("Twilight".into())
    .url("https://github.com/twilight-rs/twilight")
    .build();

let embed = EmbedBuilder::new().author(author).build()?;

pub fn color(self, color: u32) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the color.

This must be a valid hexadecimal RGB value. 0x000000 is not an acceptable value as it would be thrown out by Discord. Refer to COLOR_MAXIMUM for the maximum acceptable value.

Examples

Set the color of an embed to 0xfd69b3:

use twilight_embed_builder::EmbedBuilder;

let embed = EmbedBuilder::new()
    .color(0xfd_69_b3)
    .description("a description")
    .build()?;

pub fn description(self, description: impl Into<String>) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the description.

Refer to DESCRIPTION_LENGTH_LIMIT for the maximum number of UTF-16 code points that can be in a description.

Examples

use twilight_embed_builder::EmbedBuilder;

let embed = EmbedBuilder::new().description("this is an embed").build()?;

pub fn field(self, field: impl Into<EmbedField>) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Add a field to the embed.

Examples

use twilight_embed_builder::{EmbedBuilder, EmbedFieldBuilder};

let embed = EmbedBuilder::new()
    .description("this is an embed")
    .field(EmbedFieldBuilder::new("a field", "and its value"))
    .build()?;

pub fn footer(self, footer: impl Into<EmbedFooter>) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the footer of the embed.

Examples

use twilight_embed_builder::{EmbedBuilder, EmbedFooterBuilder};

let embed = EmbedBuilder::new()
    .description("this is an embed")
    .footer(EmbedFooterBuilder::new("a footer"))
    .build()?;

pub fn image(self, image_source: ImageSource) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the image.

Examples

Set the image source to a URL:

use twilight_embed_builder::{EmbedBuilder, EmbedFooterBuilder, ImageSource};

let source = ImageSource::url("https://raw.githubusercontent.com/twilight-rs/twilight/main/logo.png")?;
let embed = EmbedBuilder::new()
    .footer(EmbedFooterBuilder::new("twilight"))
    .image(source)
    .build()?;

pub fn thumbnail(self, image_source: ImageSource) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Add a thumbnail.

Examples

Set the thumbnail to an image attachment with the filename "twilight.png":

use twilight_embed_builder::{EmbedBuilder, ImageSource};

let embed = EmbedBuilder::new()
    .description("a picture of twilight")
    .thumbnail(ImageSource::attachment("twilight.png")?)
    .build()?;

pub const fn timestamp(self, timestamp: Timestamp) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the ISO 8601 timestamp.

pub fn title(self, title: impl Into<String>) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the title.

Refer to TITLE_LENGTH_LIMIT for the maximum number of UTF-16 code points that can be in a title.

Examples

Set the title to “twilight”:

use twilight_embed_builder::EmbedBuilder;

let embed = EmbedBuilder::new()
    .title("twilight")
    .url("https://github.com/twilight-rs/twilight")
    .build()?;

pub fn url(self, url: impl Into<String>) -> Self

👎 Deprecated since 0.10.1:
use twilight_util::builder::embed

Set the URL.

Examples

Set the URL to twilight’s repository:

use twilight_embed_builder::{EmbedBuilder, EmbedFooterBuilder};

let embed = EmbedBuilder::new()
    .description("twilight's repository")
    .url("https://github.com/twilight-rs/twilight")
    .build()?;